AZUL: Se mejoraron los test y el tipo de dato del saldo
Se creo un inicializador para los test, y se cambio el tipo de dato del saldo de float a double

+++ src/cuentaBancaria/CuentaBancaria.java
@@ -2,14 +2,14 @@ package cuentaBancaria;
 
 public class CuentaBancaria {
 	
-	private float saldo = 0;
+	private double saldo = 0;
 
-	public float getSaldo(){
+	public double getSaldo(){
 
 		return this.saldo;
 	}
 
-	public void depositar(final float nuevoSaldo) {
+	public void depositar(final double nuevoSaldo) {
 		this.saldo+= nuevoSaldo;
 	}	
 }
